Extract functional/reactions libraries into modules
Refactor: move the large monomer functional-groups and reaction definition dictionaries out of the detectors into two new modules (detectors/functional_groups_library.py and detectors/reactions_library.py). Update detectors to import these libraries (with try/except fallbacks for relative imports), add from __future__ import annotations and a module-level logger in the functional-groups detector, and clean up in-file definitions and commented blocks. This improves modularity, reduces detector file size, and makes the libraries reusable.
