+++ src/reaction_lammps_mupt/detectors/functional_groups_detector.py
@@ -96,137 +96,25 @@ For each monomer index, the detector returns:
 later to determine how many reactions are possible and how to build reaction
 templates safely.
 """
+
+from __future__ import annotations # Enable postponed evaluation of annotations for forward references.
+from dataclasses import dataclass
 from rdkit import Chem
+import logging
 import json
+try:
+    from functional_groups_library import FunctionalGroupsLibrary
+except (ImportError, ModuleNotFoundError):
+    from .functional_groups_library import FunctionalGroupsLibrary
+logger = logging.getLogger(__name__)  # Module-level logger for future diagnostics.
+
 
 # Dictionary defining various types of monomers, including their functionality types (e.g., 'vinyl', 'mono', 'di_different', 'di_identical'),
 # SMARTS patterns for substructure matching, and group names for identification.
 # Additional functional groups can be added based on references like "J. Chem. Inf. Model. 2023, 63, 5539−5548".
 # Note: Potential for monomers with mixed groups like COCl and COOH is unaddressed.
-# Can be many more functional groups added here
+# Can be many more functional groups added to the library based on literature and user needs.
 
-class FunctionalGroupsLibrary:
-    def __init__(self):
-        self.monomer_types = {
-            "hydroxy_carboxylic_acid_monomer": {
-                "functionality_type": "di_different",
-                "smarts_1": "[OX2H1;!$(OC=*):1]",
-                "smarts_2": "[CX3:2](=[O])[OX2H1]",
-                "group_name": "hydroxy_carboxylic_acid"
-            },
-            "hydroxy_acid_halides_monomer": {
-                "functionality_type": "di_different",
-                "smarts_1": "[OX2H1;!$(OC=*):1]",
-                "smarts_2": "[CX3:2](=[O])[Cl,Br,I]",
-                "group_name": "hydroxy_acid_halide",
-                "comments": "Hydroxy acid halides are highly reactive and less commonly used monomers for polyesterification compared to hydroxy carboxylic acids."
-            },  
-            "diol_monomer": {
-                "functionality_type": "di_identical",
-                "smarts_1": "[O,S;X2;H1;!$([O,S]C=*):3]",
-                "group_name": "diol",
-                "comments": None,
-            },
-
-            "amino_acid_monomer": {
-                "functionality_type": "di_different",
-                "smarts_1": "[NX3;H2,H1;!$(OC=*):1]",
-                "smarts_2": "[CX3:2](=[O])[OX2H1]",
-                "group_name": "amino_acid",
-                "comments": None,
-            },
-            "di_amine_monomer": { 
-                "functionality_type": "di_identical",
-                "smarts_1": "[N&X3;H2,H1;!$(NC=*):3]",
-                "group_name": "di_amine",
-                "comments": None,
-            },
-            "di_carboxylic_acid_monomer": {
-                "functionality_type": "di_identical",
-                "smarts_1": "[CX3:2](=[O])[OX2H1:1]",
-                "group_name": "di_carboxylic_acid",
-                "comments": None,
-            },
-            "di_carboxylic_acid_halide_monomer": {
-                "functionality_type": "di_identical",
-                "smarts_1": "[CX3:2](=[O])[Cl,Br,I:1]",
-                "group_name": "di_carboxylic_acid_halide",
-                "comments": None,
-            },
-            # "di_cyclic_anhydride_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[CX3,c;R:1](=[OX1])[OX2,o;R][CX3,c;R:2](=[OX1])",
-            #     "group_name": "di_cyclic_anhydride"
-            # },
-            # "di_isocyanate_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[NX2:1]=[CX2]=[OX1,SX1:2]",
-            #     "group_name": "di_isocyanate"
-            # },
-            # "di_epoxide_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[CX4;H2,H1,H0;R:1]1[OX2;R:2][CX4;H1,H0;R:3]1",
-            #     "group_name": "di_epoxide"
-            # }
-            # "vinyl_monomer": {
-            #     "functionality_type": "vinyl",
-            #     "smarts_1": "[C]=[C;D1]",
-            #     "group_name": "vinyl"
-            # },
-            # "cyclic_olefin_monomer": {
-            #     "functionality_type": "vinyl",
-            #     "smarts_1": "[CX3;R:1]=[CX3;R:2]",
-            #     "group_name": "cyclic_olefin"
-            # },
-            # "lactone_monomer": {
-            #     "functionality_type": "mono",
-            #     "smarts_1": "[CX3;R:1](=[OX1])[OX2;R:2]",
-            #     "group_name": "lactone"
-            # },
-            # "cyclic_anhydride_monomer": {
-            #     "functionality_type": "mono",
-            #     "smarts_1": "[C,c;R:1][CX3,c;R](=[OX1])[OX2,o;R][CX3,c;R](=[OX1])[C,c;R:2]",
-            #     "group_name": "cyclic_anhydride"
-            # },
-            # "epoxide_monomer": {
-            #     "functionality_type": "mono",
-            #     "smarts_1": "[CX4;R:3]1[OX2;R:4][CX4;R:5]1",
-            #     "group_name": "epoxide"
-            # },
-            # "lactam_monomer": {
-            #     "functionality_type": "mono",
-            #     "smarts_1": "[CX3;R:1](=[OX1])[NX3;R:2]",
-            #     "group_name": "lactam"
-            # },
-            # "di_amine_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[N&X3;H2,H1;!$(NC=*):3]",
-
-            #     "group_name": "di_amine"
-            # },
-            # "primery_di_amine_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[C,c:6][NX3;H2;!$(N[C,S]=*)]",
-            #     "group_name": "di_primery_amine"
-            # },
-            # "di_cyclic_anhydride_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[CX3,c;R:1](=[OX1])[OX2,o;R][CX3,c;R:2](=[OX1])",
-            #     "group_name": "di_cyclic_anhydride"
-            # },
-            # "di_isocyanate_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[NX2:1]=[CX2]=[OX1,SX1:2]",
-            #     "group_name": "di_isocyanate"
-            # },
-            # "di_epoxide_monomer": {
-            #     "functionality_type": "di_identical",
-            #     "smarts_1": "[CX4;H2,H1,H0;R:1]1[OX2;R:2][CX4;H1,H0;R:3]1",
-            #     "group_name": "di_epoxide"
-            # }
-            # need to add more functional groups here from "J. Chem. Inf. Model. 2023, 63, 5539−5548"
-        }
-        # is there monomers with both COCl and COOH groups?
 
 class FunctionalGroupsDetector:
     def __init__(self):
